RasberryPi/requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
paho-mqtt
